Partitioning before Merging
In order to be able to process large datasets, we need to partition
input into smaller patches so that Merging would be possible to run on a
ordinary VM. Here we add the ability to partition variants based on
thier reference_name (CHROM) value.

This change enables VT to process Platinum_100 (with a caviate of some
changes needed for writing to BigQuery table).