Making num of output splits an input flag
Following the last PR attemping to fix the BQ sink bug:
https://github.com/googlegenomics/gcp-variant-transforms/pull/226

We realized when we have both partitioning before merge and partitioning
before write_to_bq, one of our integration test fails due to larger time
that DataFlow needs to bring up the pipeline.

In order to fix thie problem we added a new flag to make it possible:
1) To set the num of output splits dynamically (as opposed to default 20)
2) Make that option independent from optimize_for_large_inputs flag.

With this change it will be possible to partition before merge and/or
partition before write.