allow incompatible schema (#370)
In BQ to VCF pipeline, an error occurs if the schema is not compatible with the reserved fields in VCF 4.3 spec. Currently there is no easy way to load these tables. Therefore, this PR:

Provide an option --allow_incompatible_schema to disregard the reserved fields.

Issues: 85
Tested: Manually ran BQ to VCF test with the option.