update the temp file name for each job to be unique (#371)
When there are multiple test cases in BQ to VCF integration tests, it happens that the temp folders and some intermediate files are using the same name for different jobs. This PR

create a unique id for the temp folders and files that are unique for each job.

Issues: 85
Tested: unit tests, manually ran bq to vcf.