Store sample ID hash hex string in BQ tables. (#542)
This PR is an initial step of changing the way we handle sample names in our tables. In future PRs, this field in our codebase and in BQ tables (currently call.name) should be renamed to call.sample_id and then stored as hash integer instead of a hex string.