Delete empty tables after AVRO copy stage
In order to make the output of VT more clear, in this PR we examine all output tables and delete empty ones.
Empty tables are created when in the sharding config we have an entry for a chromosome (say ChrM) while the input VCF file does not have any line with that reference_name.