Move the delete empty tables logic to LoadAvro class
In our tests we found out the method that was executing `SELECT COUNT(1)` query does not work reliably and sometimes we ended up deleting tables that were actually not empty. It seems this is the same issue as our flaky integration tests were sometimes tables that are expected not to be empty return 0 as the number of rows.
We are trying to address that issue in #599