Add a new sharding config with fixed partition range end values
This was a feedback by one of our users who wanted to define a view on all of their output tables. If tables have different partition range ends then that is not possible. Similarly, running a query with wildcard on table name would fail.

This sharding config would resolve that issue. The down side is higher query cost for smaller chromosomes, especially chromosome 19, 20, 21, 22, and Y.