Increate _MAX_NUM_REGIONS
from 10 to 64 to allow having a sharding config file with one shard that
contains potentially all chromosomes.

Some of our users prefer not to shart the output table, to do that they
use a sharding config that assigns all chromosomes to a single shard.
For that use case we need to allow larger than 10 regions per shard.